Update myStyle.css
A few touches here in there for a better UX/UI. Modified the pads (now buttons form divs) interaction using the ":active" selector to modifiy on click and for a more controlled and flexible customization. Refined the design colors and created a dark theme instead of the white one. Stylized the newly added footer div to match the codepen.io full-page view design language and integrated an avatar icon using the picture from my profile.