Update myScript.js
Made some good progress on the main logic of the game. Redefined the way the user inputs the data with the color pads. Defining step by step the main functions that make this game to work. After that I'll make the necessary connections between them to make the game autonomous.