Update myScript.js
Made some more good progress on the main logic of the game. Refined the timeout of the functions and the order they are triggered by user actions/inputs and the way the random sequence animates the pads. Next step is to write a function that compares the every element of the aiCollector array, that contains the random generated numbers, with every element of the userCollector array that contains the user input numbers based on the pads that are clicked.