Update myScript.js
Getting closer and closer to the finished game. The main game logic works as it should. Implemented a few mods to the interface and how is interactive based on the game's turn. (i.e. the buttons on the game interface are active only when the game is on and depending on if it's the user's turn to press the buttons). Written and defined an efficient way the game compares the randomized generated pattern with the one generated by the user, by converting all the separate values from each array to a concatenated code sequence. The last important part is to make the function that flashes the pads to work correctly all the way to the final level (20). Now when it has to do a sequence with more than 4-5 numbers it breaks down by flashing to fast or irregular.