Update myScript.js
Fixed the flasher() function. Now it works as is should. Made some mp3 and ogg files with my music maker software and added unique sound effects to all the pads. The sounds are urban dub-step drums inspired. I have to write a function for the situation when the user mistakes the pattern and that makes the game to repeat the last pattern. If strict mode is active to reset the game from the beginning when the user make a mistake.  When level 20 is reached to notify the user he won the game and to restart it.