Update myScript.js
The game is almost done. Created a few more functions; checkSequence(a, b) one for checking the last code sequence, restartGame() one to restart the game on cal and repeatSequence() one to repeat the last code sequence to the user. The single thing that's left to be addressed is when the strict mode is activated, the game will restart on the first wrong pad chosen by the user.